requirements.txt
folium==0.13.0
haversine==2.8.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.26.1
pandas==2.1.3
Pillow==9.2.0
plotly==5.18.0
streamlit==1.28.1
streamlit-folium==0.7.0

This repository contains files and script to build a company strategy dashboard
